llm_platform/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
pgvector
openai
python-dotenv
